fix: resolve all critical bugs and unify project to net9.0
- Fix AutoMapper namespaces: POS.Application.Mappers -> Drones.Application.Mappers
- Fix MedicationRequestViewModel.Weight: int? -> double
- Fix DroneController.ListDrones: [HttpPost] -> [HttpGet]
- Fix connection strings in all appsettings to use LocalDB (DronesDB)
- Unify all projects to net9.0 and consistent package versions
- Add Name property to NModel and NState entities (was missing, broke EF config)
- Re-enable NModel.cs, NState.cs, NModelConfiguration.cs, NStateConfiguration.cs
- Update FluentValidation to modern API (remove [Obsolete] AddFluentValidation)
- Update AutoMapper to 16.1.1, Swashbuckle to 6.9.0, EF Core to 9.0.4
- Remove unused Azure packages and Microsoft.Extensions.Azure using
- Add IDesignTimeDbContextFactory for EF migrations support
- Create initial EF Core migration and apply to create DronesDB in LocalDB

+++ src/Drones.Application/ViewModels/Medication/Request/MedicationRequestViewModel.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public int? Weight { get; set; }
+        public double? Weight { get; set; }
         public string? Code { get; set; }
         public string? Image { get; set; }
     }
